Seperate parsers to their own file
